support micro app coexist

+++ packages/micro-core/src/lib/micro-portal.service.ts
@@ -1,30 +1,25 @@
-import { Injectable, NgZone, ApplicationRef } from '@angular/core';
+import { Injectable, NgZone, ApplicationRef, Injector } from '@angular/core';
 import { NavigationEnd, RouterEvent, Router } from '@angular/router';
-import { AssetsLoader } from './assets-loader';
+import { AssetsLoader, AssetsLoadResult } from './assets-loader';
 import { MicroHostApplication } from './host-application';
 import { GlobalEventDispatcher } from './global-event-dispatcher';
-import { getHTMLElement } from './helpers';
+import { getHTMLElement, coerceArray } from './helpers';
+import { of, Observable, BehaviorSubject, Subject } from 'rxjs';
+import { tap, map } from 'rxjs/operators';
+import { IMicroApplication, ApplicationInfo, ApplicationOptions, SwitchModes } from './micro.class';
 
-export interface ApplicationOptions {
-    host: string | HTMLElement;
-    selector: string;
-    routerPathPrefix: string;
-    stylePathPrefix?: string;
-    styles?: string[];
-    scriptPathPrefix?: string;
-    scripts?: string[];
-}
-
-export interface ApplicationInfo {
-    name: string;
-    loaded?: boolean;
-    options: ApplicationOptions;
+export interface MicroPortalOptions {
+    preload?: boolean;
+    switchMode?: SwitchModes;
+    errorHandler: (error: Error) => void;
 }
 
 @Injectable({
     providedIn: 'root'
 })
 export class MicroPortalService {
+    private options: MicroPortalOptions;
+
     private apps: ApplicationInfo[] = [];
 
     private appsMap: { [key: string]: ApplicationOptions } = {};
@@ -35,6 +30,31 @@ export class MicroPortalService {
 
     public loadingDone: boolean;
 
+    public appLoad$ = new Subject<ApplicationInfo>();
+
+    private switchModeIsCoexist() {
+        return this.options.switchMode === SwitchModes.coexist;
+    }
+
+    private getMicroApplication(app: ApplicationInfo): IMicroApplication {
+        const appInstance = (window as any)[app.name] as { app: IMicroApplication };
+        return appInstance && appInstance.app;
+    }
+
+    private hideApplication(appInfo: ApplicationInfo) {
+        const appRootElement = document.querySelector(appInfo.options.selector);
+        if (appRootElement) {
+            appRootElement.setAttribute('style', 'display:none;');
+        }
+    }
+
+    private showApplication(appInfo: ApplicationInfo) {
+        const appRootElement = document.querySelector(appInfo.options.selector);
+        if (appRootElement) {
+            appRootElement.setAttribute('style', '');
+        }
+    }
+
     constructor(
         private assetsLoader: AssetsLoader,
         private ngZone: NgZone,
@@ -44,7 +64,22 @@ export class MicroPortalService {
     ) {
         this.hostApp.ngZone = ngZone;
         this.hostApp.router = router;
+
         this.hostApp.globalEventDispatcher = globalEventDispatcher;
+        this.options = {
+            switchMode: SwitchModes.default,
+            preload: true,
+            errorHandler: (error: Error) => {
+                console.error(error);
+            }
+        };
+    }
+
+    setOptions(options: Partial<MicroPortalOptions>) {
+        this.options = {
+            ...this.options,
+            ...options
+        };
     }
 
     registerApplication(appName: string, options: ApplicationOptions) {
@@ -58,9 +93,9 @@ export class MicroPortalService {
         this.appsMap[appName] = options;
     }
 
-    loadApp(app: ApplicationInfo) {
+    loadApp(app: ApplicationInfo): Observable<AssetsLoadResult[]> {
         if (app.loaded) {
-            return Promise.resolve([]);
+            return of([]);
         }
         let scripts = app.options.scripts;
         if (app.options.scriptPathPrefix) {
@@ -71,37 +106,56 @@ export class MicroPortalService {
         return this.assetsLoader.loadScripts(scripts);
     }
 
-    bootstrapApp(app: ApplicationInfo) {
+    bootstrapApp(appInfo: ApplicationInfo): IMicroApplication {
+        const app = this.getMicroApplication(appInfo);
+        if (app && app.bootstrap) {
+            const container = getHTMLElement(appInfo.options.host);
+            if (container) {
+                let appRootElement = container.querySelector(appInfo.options.selector);
+                if (!appRootElement) {
+                    appRootElement = document.createElement(appInfo.options.selector);
+                    if (appInfo.options.hostClass) {
+                        appRootElement.classList.add(...coerceArray(appInfo.options.hostClass));
+                    }
+                    container.appendChild(appRootElement);
+                }
+            }
+            app.bootstrap(this.hostApp);
+            return app;
+        }
+        return null;
+    }
+
+    loadAndBootstrapApp(appInfo: ApplicationInfo) {
         this.ngZone.runOutsideAngular(() => {
             this.loadingDone = false;
-            this.loadApp(app).then(result => {
-                app.loaded = true;
-                const appInstance = (window as any)[app.name];
-                this.currentApp = app;
-                if (appInstance && appInstance.app && appInstance.app.bootstrap) {
-                    const container = getHTMLElement(app.options.host);
-                    if (container) {
-                        let appRootElement = container.querySelector(app.options.selector);
-                        if (!appRootElement) {
-                            appRootElement = window.document.createElement(app.options.selector);
-                            appRootElement.setAttribute('class', 'thy-layout');
-                            container.appendChild(appRootElement);
-                        }
-                    }
-                    appInstance.app.bootstrap(this.hostApp);
+            this.currentApp = appInfo;
+            if (appInfo.loaded && this.switchModeIsCoexist()) {
+                this.showApplication(appInfo);
+                const app = this.getMicroApplication(appInfo);
+                app.resetRouting();
+                return;
+            }
+            this.loadApp(appInfo).subscribe(
+                result => {
+                    this.bootstrapApp(appInfo);
+                    appInfo.loaded = true;
                     this.loadingDone = true;
+                },
+                error => {
+                    this.options.errorHandler(error);
                 }
-            });
+            );
         });
     }
 
-    destroyApplication(app: ApplicationInfo) {
-        const appInstance = (window as any)[app.name];
-        if (appInstance) {
-            appInstance.app.destroy();
+    destroyApplication(appInfo: ApplicationInfo) {
+        const app = this.getMicroApplication(appInfo);
+        if (app) {
+            app.destroy();
         }
-        const container = getHTMLElement(app.options.host);
-        const appRootElement = container.querySelector(app.options.selector);
+        const container = getHTMLElement(appInfo.options.host);
+        const appRootElement = container.querySelector(appInfo.options.selector);
         if (appRootElement) {
             container.removeChild(appRootElement);
         }
@@ -114,13 +168,25 @@ export class MicroPortalService {
     }
 
     resetRouting(event: RouterEvent) {
-        if (this.currentApp) {
-            this.destroyApplication(this.currentApp);
-        }
-        this.apps.forEach(app => {
-            if (event.url.includes(app.options.routerPathPrefix)) {
-                this.bootstrapApp(app);
-            }
+        const matchedApp = this.apps.find(app => {
+            return event.url.includes(app.options.routerPathPrefix);
         });
+
+        if (this.currentApp) {
+            if (this.switchModeIsCoexist()) {
+                const app = this.getMicroApplication(this.currentApp);
+                if (app) {
+                    this.hideApplication(this.currentApp);
+                    app.resetRouting();
+                }
+            } else {
+                this.destroyApplication(this.currentApp);
+            }
+            this.currentApp = null;
+        }
+
+        if (matchedApp) {
+            this.loadAndBootstrapApp(matchedApp);
+        }
     }
 }
